# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Online Course Enrollment System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enrollments/pay:
    post:
      operationId: processPayment
      tags: [enrollments]
      summary: Process payment and enroll in a course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "201":
          description: Payment successful and enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          description: Bad request or already enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        paymentMethod:
          type: string
          enum: [upi, credit_card, net_banking]
        paymentStatus:
          type: string
          enum: [pending, completed, failed]
        amountPaid:
          type: number
        transactionId:
          type: string
        enrolledAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - courseId
        - paymentMethod
        - paymentStatus
        - amountPaid
        - transactionId
        - enrolledAt
...
    PaymentRequest:
      type: object
      properties:
        courseId:
          type: integer
        paymentMethod:
          type: string
          enum: [upi, credit_card, net_banking, qr_code]
        paymentDetails:
          type: object
          additionalProperties: true
      required:
        - courseId
        - paymentMethod
        - paymentDetails
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        enrollment:
          $ref: "#/components/schemas/Enrollment"
          nullable: true
        message:
          type: string
      required:
        - success
        - transactionId
        - message
...